+++ omisego-sdk/src/main/java/co/omisego/omisego/model/transaction/request/TransactionRequestCreateParams.kt
@@ -27,7 +27,7 @@ data class TransactionRequestParams(
      * The amount of token to receive
      * This amount can be either inputted when generating or consuming a transaction request.
      */
-    val amount: Double? = null,
+    val amount: Int? = null,
 
     /**
      * The address specifying where the transaction should be sent to.
